403 en gacha

+++ app/src/main/java/com/patitofeliz/fireemblem/core/model/api/PullResponse.kt
@@ -2,7 +2,7 @@ package com.patitofeliz.fireemblem.core.model.api
 
 data class PullResponse(val pullId: Int,
     val usuarioId: Int,
-    val bannerId: Int,
+    val bannerId: Long,
     val nombre: String,
     val clase: String,
     val tipo: String,
